+++ com.unity.multiplayer.mlapi/Tests/PlayMode/RPCTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 95bdda1c14fd04144b03fa1943b207fb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
