+++ com.unity.multiplayer.mlapi/Tests/PlayMode/RPCTests/NetworkingManagerTests.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 685930534aa4e8744817ceb78adbb66d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
