+++ com.unity.multiplayer.mlapi/Tests/PlayMode/RPCTests/Player.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c9ae7eb0bad342a449caef992d16eccd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
