+++ com.unity.multiplayer.mlapi/Tests/PlayMode/RPCTests/RpcTestingObject.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0b2a6e0622c6b2548b641288f1088be0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
